Minor changes to fit the codebase (bug fixes)

+++ main/headers/crc16.h
@@ -1,6 +1,7 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 // Function to compute CRC16 checksum using a lookup table
-uint16_t crc16(const uint8_t *data, uint16_t length);
+uint16_t crc16(const uint8_t *data, size_t length);
